Minor cleanups to the build process
Move the repo locations to the top of the script, and add ability
to specify specific tags to be built.  They are defaulted to the
same repositories as before and set to build master, so this should
not cause a functional change by default.

Also build a static version of zlib so we get compressed mux file
support, and add in the bmsdk so we can successfully build on
non-Linux platforms (the tree has a copy of the bmsdk headers
included, but they are Linux specific).